requirements.txt
astroid==2.1.0
autopep8==1.4.3
beautifulsoup4==4.6.3
botocore==1.12.71
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-rest-swagger==2.2.0
djangorestframework==3.9.0
docutils==0.14
httpie==1.0.2
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
pyasn1==0.4.5
pycodestyle==2.4.0
Pygments==2.3.1
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
